Use testApp everywhere
