+++ packages/app/src/cli/services/info.test.ts
@@ -3,6 +3,7 @@ import {fetchOrgAndApps, fetchOrganizations} from './dev/fetch.js'
 import {selectOrCreateApp} from './dev/select-app.js'
 import {AppInterface} from '../models/app/app.js'
 import {selectOrganizationPrompt} from '../prompts/dev.js'
+import {testApp} from '../models/app/app.test-data.js'
 import {path, session, output, store} from '@shopify/cli-kit'
 import {describe, it, expect, vi, beforeEach} from 'vitest'
 import {checkForNewVersion} from '@shopify/cli-kit/node/node-package-manager'
@@ -148,19 +149,13 @@ describe('info', () => {
 function mockApp(currentVersion = '2.2.2'): AppInterface {
   const nodeDependencies: {[key: string]: string} = {}
   nodeDependencies['@shopify/cli'] = currentVersion
-  return {
+  return testApp({
     name: 'myapp',
-    idEnvironmentVariableName: 'SHOPIFY_API_KEY',
     directory: '/',
-    packageManager: 'yarn',
     configurationPath: path.join('/', 'shopify.app.toml'),
     configuration: {
       scopes: 'my-scope',
     },
-    webs: [],
     nodeDependencies,
-    extensions: {ui: [], function: [], theme: []},
-    updateDependencies: vi.fn(),
-    hasExtensions: vi.fn(),
-  }
+  })
 }
