+++ packages/app/src/cli/utilities/extensions/configuration.test.ts
@@ -1,6 +1,7 @@
 import {extensionConfig, ExtensionConfigOptions} from './configuration.js'
 import {AppInterface} from '../../models/app/app.js'
 import {UIExtension} from '../../models/app/extensions.js'
+import {testApp} from '../../models/app/app.test-data.js'
 import {beforeEach, describe, expect, test, vi} from 'vitest'
 import {path} from '@shopify/cli-kit'
 
@@ -23,7 +24,9 @@ beforeEach(() => {
     }
   })
   vi.mock('../../models/app/app', async () => {
+    const appOriginal: any = await vi.importActual('../../models/app/app.js')
     return {
+      ...appOriginal,
       getUIExtensionRendererVersion: () => {
         return {name: 'renderer-name', version: '2.1.5'}
       },
@@ -54,21 +57,13 @@ describe('extensionConfig', () => {
       entrySourceFilePath: `${extensionRoot}/src/index.js`,
       devUUID: 'devUUID',
     }
-    const app: AppInterface = {
+    const app: AppInterface = testApp({
       name: 'myapp',
-      idEnvironmentVariableName: 'SHOPIFY_API_KEY',
       directory: appRoot,
-      packageManager: 'yarn',
       configurationPath: path.join(appRoot, 'shopify.app.toml'),
-      configuration: {
-        scopes: '',
-      },
-      webs: [],
       nodeDependencies: {},
       extensions: {ui: [extension], function: [], theme: []},
-      updateDependencies: vi.fn(),
-      hasExtensions: vi.fn(),
-    }
+    })
 
     const options: ExtensionConfigOptions = {
       app,
